moving named constant from rag handler to rag
