+++ tests/integrations/utilities/rag/rerankers/test_openai_reranker.py
@@ -1,0 +1,17 @@
+from langchain.schema import Document
+import pytest
+
+from mindsdb.integrations.utilities.rag.rerankers.reranker_compressor import OpenAIReranker
+
+
+@pytest.mark.asyncio
+async def test_openai_reranker():
+    openai_reranker = OpenAIReranker()
+    openai_reranker.model_post_init(None)
+    results = await openai_reranker.compress_documents(
+        documents=[Document(page_content="Trump declared that he likes cats more than dogs"),
+                   Document(page_content="Trump declared that he likes AI")],
+        query="Trump opinion on animals",
+    )
+    assert len(results) == 1
+    assert "cats" in results[0].page_content
